requirements.txt
colorama==0.4.4
dnspython==2.2.1
tqdm==4.64.0
